+++ jest.config.js
@@ -34,5 +34,7 @@ module.exports = {
     ".js$": "babel-jest"
   },
 
-  setupFiles: ["<rootDir>/jest.init.js"]
+  setupFiles: ["<rootDir>/jest.init.js"],
+
+  testSequencer: "<rootDir>/jest.testSequencer.js"
 };